+++ gui.cpp
@@ -355,7 +355,7 @@ void ChooseMenPanel::draw() {
 #if SDL_MAJOR_VERSION == 1
         screen->fillRect(0, 0, default_width_c*scale_width, default_height_c*scale_height, 0, 0, 0);
 #else
-        screen->fillRectWithAlpha(0, 0, default_width_c*scale_width, default_height_c*scale_height, 0, 0, 0, 255-32);
+        screen->fillRectWithAlpha(0, 0, (short)(default_width_c*scale_width), (short)(default_height_c*scale_height), 0, 0, 0, 255-32);
 #endif
     }
 
@@ -373,7 +373,7 @@ void ChooseMenPanel::draw() {
 #if SDL_MAJOR_VERSION == 1
         screen->fillRect(0, 0, default_width_c*scale_width, default_height_c*scale_height, 0, 0, 0);
 #else
-        screen->fillRectWithAlpha(0, 0, default_width_c*scale_width, default_height_c*scale_height, 0, 0, 0, 255-32);
+        screen->fillRectWithAlpha(0, 0, (short)(default_width_c*scale_width), (short)(default_height_c*scale_height), 0, 0, 0, 255-32);
 #endif
     }
